migrate currency utils
