+++ plugins/cost-insights/src/components/CostGrowth/CostGrowth.test.tsx
@@ -18,13 +18,12 @@ import React, { PropsWithChildren } from 'react';
 import { renderInTestApp } from '@backstage/test-utils';
 import { CostGrowth } from './CostGrowth';
 import {
-  defaultCurrencies,
   Currency,
-  CurrencyType,
   Duration,
   findAlways,
 } from '../../types';
 import { MockConfigProvider, MockCurrencyProvider } from '../../utils/tests';
+import { defaultCurrencies, CurrencyType } from '../../utils/currency';
 
 const engineers = findAlways(defaultCurrencies, c => c.kind === null);
 const usd = findAlways(defaultCurrencies, c => c.kind === CurrencyType.USD);
@@ -41,10 +40,10 @@ const MockContext = ({
   currency: Currency;
   engineerCost: number;
 }>) => (
-  <MockConfigProvider engineerCost={engineerCost}>
-    <MockCurrencyProvider currency={currency}>{children}</MockCurrencyProvider>
-  </MockConfigProvider>
-);
+    <MockConfigProvider engineerCost={engineerCost}>
+      <MockCurrencyProvider currency={currency}>{children}</MockCurrencyProvider>
+    </MockConfigProvider>
+  );
 
 describe.each`
   engineerCost | ratio           | amount     | expected
