+++ plugins/cost-insights/src/types/Currency.ts
@@ -13,15 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Duration } from './Duration';
-import { assertNever } from './Maybe';
-
-export enum CurrencyType {
-  USD = 'USD',
-  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
-  Beers = 'BEERS',
-  IceCream = 'PINTS_OF_ICE_CREAM',
-}
 
 export interface Currency {
   kind: string | null;
@@ -30,49 +21,3 @@ export interface Currency {
   prefix?: string;
   rate?: number;
 }
-
-export const rateOf = (cost: number, duration: Duration) => {
-  switch (duration) {
-    case Duration.P1M:
-    case Duration.P30D:
-      return cost / 12;
-    case Duration.P90D:
-    case Duration.P3M:
-      return cost / 4;
-    default:
-      return assertNever(duration);
-  }
-};
-
-export const defaultCurrencies: Currency[] = [
-  {
-    kind: null,
-    label: 'Engineers 🛠',
-    unit: 'engineer',
-  },
-  {
-    kind: CurrencyType.USD,
-    label: 'US Dollars 💵',
-    unit: 'dollar',
-    prefix: '$',
-    rate: 1,
-  },
-  {
-    kind: CurrencyType.CarbonOffsetTons,
-    label: 'Carbon Offset Tons ♻️⚖️s',
-    unit: 'carbon offset ton',
-    rate: 3.5,
-  },
-  {
-    kind: CurrencyType.Beers,
-    label: 'Beers 🍺',
-    unit: 'beer',
-    rate: 4.5,
-  },
-  {
-    kind: CurrencyType.IceCream,
-    label: 'Pints of Ice Cream 🍦',
-    unit: 'ice cream pint',
-    rate: 5.5,
-  },
-];
